refactor(mlx): clear the clang-tidy findings clang-apply-replacements missed
The first sweep (6af4b1c7) auto-applied ~554 of 841; clang-apply-
replacements drops fixes on macro expansions, multi-variable declarations,
and (a dedup miss) shared headers, leaving 287. Plus the now-uncovered
clang-analyzer tier (it only runs once the surface findings clear).

RED before this commit: full `lint-c-mlx` reported 287 then, after the
high-leverage fixes, 9 clang-analyzer findings.

High-leverage one-edit clears:
- WITH_STREAM macro: `const mx::StreamContext _stream_guard` (162
  misc-const-correctness on the RAII guard, one per streamed op).
- tape.h OP_* enum: `enum : std::uint8_t` (70 opcodes; 74
  performance-enum-size, one per TU).
- Tensor ctor param `rg` -> `requires_grad` (8 bugprone-argument-comment:
  the `/*requires_grad=*/` call-site comments now match).

Targeted fixes:
- 25 combined-declaration `int X = .., Y = ..` -> `int const ..` in the
  conv/pool/shape ops (clang-tidy flags one var but the whole decl is
  const-eligible).
- 2 nested dtype ternaries (batch, cast) -> switch.
- 2 size_t->int narrowings (optimizer get_m/get_v): explicit cast.
- backward.cpp OP_NAMES: `const char* const` (was `const char const*`,
  duplicate-decl-specifier + pointee-const).
- cosine_similarity: drop dead `n` (deadcode).
- diagnostics atoi -> strtol (cert-err34-c).

clang-analyzer FPs NOLINT'd with rationale (5 uninitialized.Assign on the
host<->device double-buffer marshalling helpers where the analyzer can't
prove the caller-provided buffer is initialized; 2 NonNullParamChecker on
the gather/scatter index tensor that tape_append always sets).

.clang-tidy: deny-list bugprone-branch-clone — FALSE POSITIVE on this
codebase's dtype dispatch (treats `case N: dt = <distinct Dtype>; break;`
arms as clones) and on if/else differing by one statement; every hit
(device select, 4 dtag->Dtype switches, RMSprop momentum branch) was
confirmed distinct.

Verified: `make BACKEND=mlx C_LINT_FULL_CLANG_TIDY=1 lint-c-mlx` 0
findings; mlx backend builds clean; test-unit-c 260/260; check-fmt-c
clean. mlx step stays advisory for one CI run to confirm Linux nix-21 ==
macOS nix-21 before the follow-up enforces.

+++ packages/backends/backend_mlx/training/optimizer.cpp
@@ -364,9 +364,6 @@ extern "C" void optimizer_step(OptimizerHandle h) {
 			opt->v_bufs[i] = mx::add(mx::multiply(alpha_arr, opt->v_bufs[i]),
 			                         mx::multiply(one_m_alpha, mx::square(g)));
 			auto avg = mx::add(mx::sqrt(opt->v_bufs[i]), eps_arr);
-			// Branches differ (then updates m_bufs then t->data; else
-			// updates t->data directly); clang-tidy branch-clone FP.
-			// NOLINTNEXTLINE(bugprone-branch-clone)
 			if (opt->momentum > 0) {
 				opt->m_bufs[i] =
 				    mx::add(mx::multiply(momentum_a, opt->m_bufs[i]), mx::divide(g, avg));
@@ -515,7 +512,7 @@ extern "C" int optimizer_buf_count(OptimizerHandle h) {
 extern "C" void optimizer_get_m(OptimizerHandle h, int idx, double* out) {
 	auto* opt = (Optimizer*)h;
 	if (idx >= (int)opt->m_bufs.size()) {
-		int const n = ((Tensor*)param_tensor(idx))->data.size();
+		int const n = (int)((Tensor*)param_tensor(idx))->data.size();
 		memset(out, 0, n * sizeof(double));
 		return;
 	}
@@ -527,7 +524,7 @@ extern "C" void optimizer_get_m(OptimizerHandle h, int idx, double* out) {
 extern "C" void optimizer_get_v(OptimizerHandle h, int idx, double* out) {
 	auto* opt = (Optimizer*)h;
 	if (idx >= (int)opt->v_bufs.size()) {
-		int const n = ((Tensor*)param_tensor(idx))->data.size();
+		int const n = (int)((Tensor*)param_tensor(idx))->data.size();
 		memset(out, 0, n * sizeof(double));
 		return;
 	}
